Finalziación del primer step del ejercico 10 de la práctica 5.  Comienzo del segundo step.

+++ JuanCarlosMunoz/springbatchf/src/main/java/com/viewnext/springbatchf/model/entity/Order.java
@@ -4,9 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-
-import javax.persistence.Id;
 
 @Document
 @Data
@@ -18,6 +17,6 @@ public class Order {
     @Id
     private Long numOrder;
     private User client;
-    private float price;
+    private Double price;
 
 }
